Split states and environments for propositions
States change explicitly over time, while environments
are tied directly to the syntactic structure of the proposition.

The environment will always shadow the state: in case of name
clashes, the variable from the environment will be used.

This is necessary to implement an evaluator function
that evaluates a proposition for a full trace.