Rename project and modules
Since we use StreamData instead of PropCheck, we no longer use
the prefix Prop. Instead, we call it QuickLTL, as it will eventually
be an implementation of (a dialiect of) this logic, as proposed in
the Quickstrom paper.

The modules were renamed accordingly and restructured.
We now have the implementation of the logic by itself as
a top-level module hierarchy. Test machinery is built on top
of it, but in separate modules.