Fix stepwise evaluation of QuickLTL propositions
The previous implementation was broken, since simplify/1
was incorrect. The main issue is that strong and weak until
are not duals. Instead, I was missing the strong and weak
release operators.

I have therefore introduced these operators and fixed the code,
as evidenced by property tests.

I don't think, however, that these operators are good for usability,
as `release_X(p, q)` is exactly the same as `until_X(p and q, q)`.
These additional operators with such subtle semantics are bound to
introduce confusion.

Thus, I should probably make them only part of the internal
representation in the future, never exposing them to users of this
library.